// requirements/enums/ActivityTypeEnum.ts
export enum ActivityTypeEnum {
  Alkoholkonsum = "Alkoholkonsum",
  Ausgaben = "Ausgaben",
  Beziehungszeit = "Beziehungszeit (Quality Time)",
  Bildschirmzeit = "Bildschirmzeit",
  Drogenkonsum = "Drogenkonsum",
  Energie = "Energie",
  Ernaehrung = "Ernährung",
  Familienzeit = "Familienzeit",
  Getraenkekonsum = "Getränkekonsum",
  Gewohnheitsbruch = "Gewohnheitsbruch (Rückfälle)",
  Haushaltsaktivitaeten = "Haushaltsaktivitaeten",
  Joggen = "Joggen",
  Koffeinkonsum = "Koffeinkonsum",
  Kreativitaet = "Kreativitaet",
  Kunst = "Kunst",
  Laufen = "Laufen",
  Lernen = "Lernen",
  Lesen = "Lesen",
  Meditation = "Meditation",
  Motivation = "Motivation",
  Musikpraxis = "Musikpraxis",
  Natur = "Natur",
  Produktivitaet = "Produktivität",
  Radfahren = "Radfahren",
  Rauchverhalten = "Rauchverhalten",
  Schlaf = "Schlaf",
  Schreiben = "Schreiben",
  Schwimmen = "Schwimmen",
  SozialeInteraktion = "Soziale Interaktion",
  Sparen = "Sparen",
  Spazieren = "Spazieren",
  Sprachtraining = "Sprachtraining",
  Stimmung = "Stimmung",
  Stress = "Stress",
  Wandern = "Wandern",
  Wasserkonsum = "Wasserkonsum",
  Workout = "Workout",
  Yoga = "Yoga",  
}