// requirements/enums/FieldKeyEnum.ts
export enum FieldKeyEnum {
  AlcoholType = "alcohol_type",
  Altitude = "altitude",
  AnxietyLevel = "anxiety_level",
  BreakDuration = "break_duration",
  CaloryConsumption = "calory_consumption",
  ConcentrationLevel = "concentration_level",
  ConflictLevel = "conflict_level",
  ConsumptionCount = "consumption_count",
  ConsumptionMethod = "consumption_method",
  ConsumptionProduct = "Consumption_product",
  CostsAmount = "costs_amount",
  CravingIntensity = "craving_intensity",
  DeviceMainUse = "device_main_use",
  DeviceType = "device_type",
  Distance = "distance",
  DrinkingAmount = "drinking_amount",
  DrinkingTemperature = "drinking_temperature",
  DrinkingType = "drinking_type",
  Duration = "duration",
  EatingContext = "eating_context",
  ExerciseDuration = "exercise_duration",
  FocusLevel = "focus_level",
  FoodQuality = "food_quality",
  HabitType = "habit_type",
  HarmonyLevel = "harmony_level",
  Heartbeat = "heartbeat",
  HouseArea = "house_area",
  HouseholdTask = "household_task",
  HungerLevel = "hunger_level",
  Initiator = "initiator",
  IntensityLevel = "intensity_level",
  Language = "language",
  LanguageTrainingType = "language_training_type",
  LearnFormat = "learn_format",
  LocationType = "location_type",
  MainMood = "main_mood",
  MealType = "meal_type",
  MentalEnergy = "mental_energy",
  MoodAfter = "mood_after",
  MoodBefore = "mood_before",
  MoodLevel = "mood_level",
  MotivationReference = "motivation_reference",
  MotivationLevel = "motivation_level",
  MovementIntensity = "movement_intensity",
  Notes = "notes",
  NumberOfBreaks = "number_of_breaks",
  NumberOfExercises = "number_of_exercises",
  NumberOfMembers = "number_of_members",
  NumberOfPages = "number_of_pages",
  NumberOfSteps = "number_of_steps",
  NumberOfWakeup = "number_of_wakeup",
  Obstacle = "obstacle",
  Occasion = "occasion",
  PhysicalDiscomfort = "physical_discomfort",
  PhysicalEnergy = "physical_energy",
  PortionSize = "portion_size",
  QualityLevel = "quality_level",
  ReadingVelocity = "reading_velocity",
  ReadMedium = "read_medium",
  RouteType = "route_type",
  SatietyLevel = "satiety_level",
  SatisfactionLevel = "satisfaction_level",
  SelfConfidence = "self_confidence",
  SideEffect = "side_effects",
  SleepDuration = "sleep_duration",
  SleepQuality = "sleep_quality",
  SocialContext = "social_context",
  Status = "status",
  StressLevel = "stress_level",
  SubstanceType = "substance_type",
  SuccessLevel = "success_level",
  SwimmingStyle = "swimming_style",
  TimeOfDay = "time_of_day",
  Trigger = "trigger",
  WaterSource = "water_source",
  WaterTemperature = "water_temperature",
  Weather = "weather",
  WorkoutType = "workout_type",
  WriteMedium = "write_medium",
}